change landing page and migrate create user logic
