+++ src/app/sign-in/[[...sign-in]]/page.tsx
@@ -3,7 +3,7 @@ import { SignIn } from '@clerk/nextjs';
 export default function Page() {
   return (
     <div className='flex items-center justify-center min-h-screen'>
-      <SignIn />
+      <SignIn afterSignInUrl={'/home'} />
     </div>
   );
 }
