+++ src/app/sign-up/[[...sign-up]]/page.tsx
@@ -3,7 +3,7 @@ import { SignUp } from '@clerk/nextjs';
 export default function Page() {
   return (
     <div className='flex items-center justify-center min-h-screen'>
-      <SignUp afterSignUpUrl={'/'} />
+      <SignUp afterSignUpUrl={'/home'} />
     </div>
   );
 }
